refactor(java): base tokens on tape class

+++ java/parser/src/main/java/com/github/razorapid/morpheus/lang/parser/Parser.java
@@ -1010,11 +1010,11 @@ public class Parser {
         return peekToken();
     }
 
-    long mark() {
+    int mark() {
         return tokens.mark();
     }
 
-    void restore(long pos) {
+    void restore(int pos) {
         tokens.restore(pos);
     }
